Improve awesome/git-repo-age error message
The "Awesome list must reside in a valid git repository" error message
(awesome/git-repo-age) also appears when (e.g. Travis) makes a shallow
copy of the awesome repo.  This is documented in the README file, but
users may miss it.  Rephrase error message and add a pointer to the
README file.